renamed type module to avoid clash with python builtin
Fixes #1624
