+++ python/hail/__init__.py
@@ -1,10 +1,10 @@
+from hail.representation import IntervalTree
+
+import hail.expr
 from hail.context import HailContext
 from hail.dataset import VariantDataset
+from hail.expr import Type
 from hail.keytable import KeyTable
 from hail.utils import TextTableConfig
-from hail.type import Type
-import hail.representation
-from hail.representation import IntervalTree
-import hail.type
 
-__all__ = ['HailContext', 'VariantDataset', 'KeyTable', 'TextTableConfig', 'IntervalTree', 'type', 'representation']
+__all__ = ['HailContext', 'VariantDataset', 'KeyTable', 'TextTableConfig', 'IntervalTree', 'expr', 'representation']
